requirements.txt
beautifulsoup4
requests
pandas
numpy
faiss-cpu
Flask
openai
python-dotenv
